Stop forcing jumpstamp nops in ngen images
We now rely upon the common code in the jitinterface CEEInfo::GetBaseCompileFlags()
to set the flag CORJIT_FLAG_PROF_REJIT_NOPS for both the JIT and Ngen cases.

Added comments whenever the flag CORJIT_FLAG_PROF_REJIT_NOPS is used to clarify
that it only controls the jumpstamp variant of re-jittable code as we now have an
alternative way to perform most re-jits.